Removed incorrect depth comparison.
From gfx-rs/wgpu#1841, the summary is that this was essentially no-ops
before. I was requesting depth comparison but I never used a depth
comparison sampler within the actual shader.

This change removes the depth testing from the sampler, which is used to
read from the texture and has no effect on output zdepth.